refactor net utils
