+++ utils/__init__.py
@@ -1,5 +1,6 @@
 from .argparser import *
 from .defines import *
-from .methods import *
+from .os_utils import *
+from .net_utils import *
 from .output_manager import *
